+++ matrx/actions/__init__.py
@@ -2,3 +2,4 @@ from .action import *
 from .door_actions import *
 from .move_actions import *
 from .object_actions import *
+
