+++ matrx/agents/__init__.py
@@ -1,0 +1,5 @@
+from .capabilities.capability import *
+from .agent_brain import *
+from .human_agent_brain import *
+from .patrolling_agent import *
+
