+++ matrx/agents/agent_brain.py
@@ -1,10 +1,13 @@
 import copy
 from typing import Union
 import random
+import numpy as np
+from matrx.actions import GrabObject, OpenDoorAction, CloseDoorAction, RemoveObject
+from matrx.utils import Message
 
-from matrx.actions.door_actions import *
-from matrx.actions.object_actions import *
-from matrx.utils.message import Message
+def func1():
+    pass
+
 
 class AgentBrain:
