--- matrx/agents/capabilities/__init__.py
@@ -1,1 +1,0 @@
-from .capability import *
