+++ matrx/objects/__init__.py
@@ -1,4 +1,4 @@
-from .simple_objects import *
-from .advanced_objects import *
-from .agent_body import *
 from .env_object import *
+from .standard_objects import *
+from .agent_body import *
+
